compute: factor out `sink::correction` module
This commit moves the `UpdateStash` and its internals to a new
`correction` module, to remove some noise from the `persist_sink`
operator implementation code. It also renames `UpdateStash` to
`Correction`, to make it clear that this is not a generic collection
type but one purpose-built for `persist_sink`.
